+++ packages/sdk-deno/mod_test.ts
@@ -3,8 +3,8 @@ import { assertEquals } from "./test_deps.ts";
 
 Deno.test("test comments.query", async () => {
 
-    let convo = new Convo('CSCpPwHnkB3niBJiUjy92YGP6xVkVZbWfK8xriDO');
-    let comments = await convo.comments.query({threadId: 'KIGZUnR4RzXDFheXoOwo', page: 1, pageSize: 10});
+    const convo = new Convo('CSCpPwHnkB3niBJiUjy92YGP6xVkVZbWfK8xriDO');
+    const comments = await convo.comments.query({threadId: 'KIGZUnR4RzXDFheXoOwo', page: 1, pageSize: 10});
     assertEquals(comments.length, 10);
 
 });