+++ packages/sdk-deno/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 0.1.1
+
+- Start Support for `convo.omnid`
+  - `convo.omnid.computeTrustScore` coming soon.
+
 ## 0.1.0
-- Release Deno Support for `convo.comments`, `convo.threads`, `convo.auth`, `convo.utils`.
-    -   `convo.omnid`, `convo.auth.getSignatureDataV2`, `convo.auth.parseSignatureV2` coming soon.
+
+- Release Deno Support for `convo.comments`, `convo.threads`, `convo.auth`,
+  `convo.utils`.
+  - `convo.omnid`, `convo.auth.getSignatureDataV2`,
+    `convo.auth.parseSignatureV2` coming soon.